refactor: InMemoryCartRepository 삭제 및 테스트 삭제
