+++ README.md
@@ -5,7 +5,7 @@
 ### 도메인 및 데이터 계층
 - [x] 상품별 장바구니 수량 상태를 관리할 수 있도록 `Cart`, `CartItem`, `CartRepository` 구조 개선
 - [x] 상품 추가/감소/삭제에 대한 상태 변경 로직 구현
-- [x] `InMemoryCartRepository`를 통해 상품 목록 화면, 상품 상세 화면, 장바구니 화면이 동일한 장바구니 상태를 공유하도록 데이터 흐름 설계
+- [x] 상품 목록 화면, 상품 상세 화면, 장바구니 화면이 동일한 장바구니 상태를 공유하도록 데이터 흐름 설계
 - [x] 시스템 프로세스 종료 이후에도 장바구니 수량 상태를 복구할 수 있도록 상태 저장 처리 적용
 - [x] 추후 Local DB 및 Server 환경으로 확장 가능하도록 Repository 인터페이스 유지
